% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_button.R
\name{action_button}
\alias{action_button}
\title{Simple wrapper around actionButton with softui CSS classes}
\usage{
action_button(
  ...,
  class = "",
  status = c("primary", "secondary", "info", "success", "danger", "warning", "light",
    "dark")
)
}
\description{
Simple wrapper around actionButton with softui CSS classes
}
