% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert}
\alias{alert}
\title{Alert boxes}
\usage{
alert(
  text,
  status = c("primary", "secondary", "info", "success", "danger", "warning", "light",
    "dark"),
  closeable = TRUE
)
}
\description{
Alert boxes
}
\details{
Does not replace `shiny_toastr`'s, just for making closable boxes withn important message
or warning. Use within `softui::box` for example.
}
