% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_box}
\alias{alert_box}
\title{Alert boxes}
\usage{
alert_box(text, status = "warning")
}
\arguments{
\item{text}{Text for the box (bold, white)}

\item{status}{Status color (see `?valid_statuses`)}
}
\description{
Alert boxes
}
\details{
Does not replace `shiny_toastr`'s, this function just makes colored boxes with an important message
or warning. Can be used as a box with `softui::box`. This box has no title, only text. 
Do not use for large amount of text, typically a one-liner with an statement that has to be read by the user.
Text color is bold and white by default (and cannot be changed).
}
