% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appInfoModule.R
\name{appInfoUI}
\alias{appInfoUI}
\alias{appInfoModule}
\title{Content for dropdown with app information}
\usage{
appInfoUI(id, info = NULL)

appInfoModule(input, output, session)
}
\arguments{
\item{id}{Shiny input id}

\item{info}{Optional - if NULL an object returned by \code{\link{read_application_info}}.}
}
\description{
Content for dropdown with app information
}
\details{
A shiny module to generate the dropdown with application information such as
publication date, git version, name, etc.
}
