% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge}
\alias{badge}
\title{SoftUI badges}
\usage{
badge(text, status = "info")
}
\arguments{
\item{text}{Text for the badge (only short text field is recommended!)}

\item{status}{Status color (see `?valid_statuses`)}
}
\description{
SoftUI badges
}
\details{
Short label (badge) to be placed before or above text, for example `badge("Informatie", status = "info")`
Font is always uppercase.
}
