% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box}
\alias{box}
\title{Content box}
\usage{
box(
  ...,
  width = 12,
  title = NULL,
  subtitle = NULL,
  icon = NULL,
  tag = "",
  tag_status = "primary",
  collapsible = TRUE,
  collapsed = FALSE,
  header_ui = NULL,
  height = NULL,
  margin_bottom = 32,
  class = ""
)
}
\arguments{
\item{width}{Bootstrap width}

\item{title}{Title for the box}

\item{subtitle}{Subtitle for the box}

\item{icon}{Optional icon (will be placed left of title)}

\item{tag}{Short label text above the title (in CAPS)}

\item{header_ui}{More UI to place just under the title}

\item{height}{Height of the box (NULL to autosize) (400 or "400px")}
}
\description{
Content box
}
