% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_collection.R
\name{box_collection}
\alias{box_collection}
\title{@rdname box_collection
 @param width_nav Width of the menu on the left (bootstrap units)
 @param width_boxes Width of the boxes on the right
 @export}
\usage{
box_collection(
  ...,
  id = NULL,
  selected = NULL,
  title = NULL,
  width_nav = 3,
  width_boxes = 9,
  fill = FALSE,
  style = ""
)
}
\description{
@rdname box_collection
 @param width_nav Width of the menu on the left (bootstrap units)
 @param width_boxes Width of the boxes on the right
 @export
}
