% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_collection.R
\name{box_panel}
\alias{box_panel}
\title{Indexed boxes
 @description Works like a tab_box, except a menu appears on the left (in a box), 
 and content appears in boxes on the right.
 @rdname box_collection
 @export
 @examples
 ui <- softui::simple_page(
box_collection(}
\usage{
box_panel(
  title = "",
  ...,
  value = title,
  icon = NULL,
  style = "",
  width = NULL
)
}
\description{
box_panel(title = "Box 1", tags$p("Dit is de eerste box met content")),
 box_panel(title = "Box 2", tags$p("Dit is de tweede box met content")),
 box_panel(title = "Box 3", tags$p("Dit is de derde box met content")),
 box_panel(title = "Box 4", tags$p("Dit is de vierde box met content"))
 
)
}
\details{
)

shinyApp(ui, function(input,output){})
}
