% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsicon.R
\name{bsicon}
\alias{bsicon}
\title{Bootstrap Icons}
\usage{
bsicon(iconname, class = NULL, status = NULL, ...)
}
\arguments{
\item{iconname}{Name of the Bootstrap icon \href{https://icons.getbootstrap.com/}{}}

\item{class}{Extra CSS class for the button (for color, use \code{status} instead)}

\item{status}{Status color (see \code{?valid_statuses})}

\item{\dots}{Extra arguments to the \code{i} tag (will become attributes of the tag)}
}
\description{
Use \code{bsicon} instead of \code{shiny::icon} in a softui application.
If status color is set, the icon is colored (with a gradient color).
}
