% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_items.R
\name{dashboard_body}
\alias{dashboard_body}
\alias{tab_items}
\alias{tab_item}
\title{Dashboard body}
\usage{
dashboard_body(...)

tab_items(..., .list = NULL)

tab_item(tabName, ...)
}
\description{
Dashboard body
}
