% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_image_header.R
\name{dashboard_header_image}
\alias{dashboard_header_image}
\alias{header_item}
\alias{dropdown_item_color}
\title{The dashboard header}
\usage{
dashboard_header_image(..., height = 150, image = NULL, icon_color = "white")

header_item(...)

dropdown_item_color(
  icon,
  ...,
  header = NULL,
  align_right = FALSE,
  icon_color = "white"
)
}
\arguments{
\item{\dots}{Any header items, usually made with \code{softui::dropdown_item}}

\item{height}{Height of the image}

\item{image}{The reference to an image in the www folder that will be the header image}

\item{icon_color}{can be any bootstrap color (primary, secondary, success, danger, warning, info, light, dark, muted and white). Defaults to white.}
}
\description{
The dashboard header

UI to place in the header bar (to the right)

Dropdown item in the header bar
}
