% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_page.R
\name{dashboard_page}
\alias{dashboard_page}
\title{Soft UI dashboard page}
\usage{
dashboard_page(
  header,
  sidebar,
  body,
  title = "",
  icon = bsicon("speedometer"),
  loadingscreen_time = 4,
  busybar_color = "#346E6E",
  disconnect_message = "De sessie is verlopen of er ging iets fout!"
)
}
\description{
Soft UI dashboard page
}
