% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable_links.R
\name{disable_link}
\alias{disable_link}
\alias{enable_link}
\alias{toggle_link}
\title{Disable/enable menu item links}
\usage{
disable_link(name)

enable_link(name)

toggle_link(name, condition = NULL)
}
\description{
Disable/enable menu item links
}
