% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect.R
\name{disconnect_message}
\alias{disconnect_message}
\title{A friendly disconnect message}
\usage{
disconnect_message(
  msg = "Oeps, er ging iets fout! Herlaad de pagina en probeer het nog een keer."
)
}
\description{
A friendly disconnect message
}
\details{
A wrapper around shinydisconnect::disconnectMessage with Shinto's settings.
}
