% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluid_page.R
\name{fluid_page}
\alias{fluid_page}
\alias{fluid_row}
\title{Content container}
\usage{
fluid_page(...)

fluid_row(..., margin = 4)
}
\arguments{
\item{margin}{Margin units, space between boxes (4 is default)}
}
\description{
Content container
}
