% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appInfoModule.R
\name{get_navigator_info}
\alias{get_navigator_info}
\title{Read navigator information and make available as a Shiny input}
\usage{
get_navigator_info(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{The session object.}
}
\description{
When run, input$navigatorInfo is created, a list with lots of info on your system, including
Shiny session clientdata (user, url, port, etc.), navigator info (window size, screen size, resolution), and browser
info (using bowser.js), version, and platform (win/mac/mobile etc.).
}
