% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user}
\alias{get_user}
\title{Get user in this session}
\usage{
get_user(default = "demo", session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{default}{If no username stored in session object, use this name (local dev)}

\item{session}{The session object (don't have to set this)}
}
\value{
Authenticated userid, character.
}
\description{
Get user in this session
}
