% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sidebar.R
\name{go_to_page}
\alias{go_to_page}
\title{Go to other 'page' (main menu item)}
\usage{
go_to_page(selected, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{selected}{The page to go to, corresponds to `tabName` in `menu_item` in the main UI}
}
\description{
Go to other 'page' (main menu item)
}
\details{
Replaces 'shinydashboard::updateTabItems', and works from within modules also.
}
