% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_action_button.R
\name{icon_action_button}
\alias{icon_action_button}
\title{Icon only button}
\usage{
icon_action_button(id, icon, status = "success", class = "", ...)
}
\arguments{
\item{id}{The shiny input id}

\item{icon}{A \code{softui::bsicon} or \code{shiny::icon}.}

\item{status}{The status color (see \code{?valid_statuses}).}

\item{\dots}{Additional arguments to \code{shiny::actionButton}}
}
\description{
A 'softui' round button with only an icon. Simple wrapper around \code{shiny::actionButton}.
Uses various softui CSS classes for styling.
}
\examples{

ui <- softui::simple_page(
  
  softui::box(width = 4,
              title = "iconActionButton",
              
              iconActionButton("id2", icon = bsicon("exclamation-circle-fill"), 
                               status = "primary"),
              shiny::tags$br(),
              iconActionButton("id1", icon = bsicon("award-fill"), status = "success"),
              shiny::tags$br(),
              iconActionButton("id2", icon = bsicon("stop-circle-fill"), 
                               status = "danger", class = "btn-lg")
  )
  
  
)
shinyApp(ui,server = function(input,output){})
}
