% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconRadioInput.R
\name{iconRadioInput}
\alias{iconRadioInput}
\title{Input modality like radio buttons-functionality with icons in the circles}
\usage{
iconRadioInput(inputId, choices, labels, icons, default_choice = NULL)
}
\arguments{
\item{inputId}{input ID for element so shiny can distinguish between them}

\item{choices}{Vector with choice-names}

\item{labels}{Vector with labels for beneath the icon-buttons}

\item{icons}{Vector with bootstrap icons so \code{softui::bsicon} can call them}

\item{default_choice}{Optional default choice which is selected initially}
}
\description{
Radio buttons-functionality with icons in the circles
}
