% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingscreen.R
\name{loadingscreen}
\alias{loadingscreen}
\title{A loading screen for Shinto applications}
\usage{
loadingscreen(time = 8)
}
\arguments{
\item{time}{Loading time (seconds)}
}
\description{
A loading screen for Shinto applications
}
