% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_item.R
\name{menu_item}
\alias{menu_item}
\alias{menu_subitem}
\alias{menu_link}
\title{Sidebar menu items}
\usage{
menu_item(text, ..., icon = bsicon("bar-chart-fill"), tabName = NULL)

menu_subitem(text, tabName)

menu_link(tabName, icon, text)
}
\description{
Sidebar menu items
}
\details{
Functions to make menu items in the sidebar menu. See `?sidebar_menu`
}
