% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalize.R
\name{modalize}
\alias{modalize}
\title{Modalize a shiny module}
\usage{
modalize(
  trigger_open = reactive(NULL),
  ui_module,
  ui_pars = list(),
  server_module,
  server_pars = list(),
  header_ui = NULL,
  footer_ui = NULL,
  ...
)
}
\arguments{
\item{trigger_open}{A reactive used to trigger the modal (often: \code{reactive(input$btn_open_modal)})}

\item{ui_module}{The module UI function}

\item{ui_pars}{A list of parameters to pass to the UI function}

\item{server_module}{The module server function}

\item{server_pars}{A list of arguments to send to the server module function}

\item{header_ui}{Extra UI to place above the module UI (in the modal)}

\item{footer_ui}{Extra UI to place below the module UI (in the modal)}

\item{\dots}{Further arguments passed to \code{\link[=modal]{modal()}} (such as \code{title = "Title!"} or \code{size = "fullscreen"})}
}
\description{
Place the UI of a module in a \code{\link[=modal]{modal()}}, and call the
server module in one command. The idea is that a module does not normally
open a modal for you, but if you need the module in a modal, use \code{modalize}.
See \code{shintocatman} package for usage.
}
