% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picker_select.R
\name{picker_select_ui}
\alias{picker_select_ui}
\alias{picker_select_module}
\title{Picker-select}
\usage{
picker_select_ui(
  id,
  label = "",
  btn_class = "btn-info",
  height = 300,
  width = 300,
  button_style = "",
  dropdown_style = ""
)

picker_select_module(
  input,
  output,
  session,
  choices = NULL,
  update = reactive(NULL),
  max_choices = 50,
  debounce = 500
)
}
\description{
Replacement for `shinyWidgets::pickerInput`, which is somehow not compatible with this (bootstrap 5.1 / softui) framework.
}
