% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_header.R
\name{populate_header}
\alias{populate_header}
\title{Write header information}
\usage{
populate_header(username)
}
\arguments{
\item{username}{Current username (for logout menu)}
}
\description{
Writes logout menu, app info menu, etc.
}
