% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressbar.R
\name{progressbar}
\alias{progressbar}
\title{Progress bar}
\usage{
progressbar(value, min = 0, max = 100, status = "info", ...)
}
\arguments{
\item{value}{Value of the bar}

\item{min}{Minimum value (default 0)}

\item{max}{Maximum value (default 100)}

\item{status}{Status color, see [valid_statuses()]}

\item{\dots}{Further arguments to the containing div, for example `style = "padding:30px;`}
}
\description{
Progress bar
}
\examples{
library(shiny)
library(softui)

ui <- softui::simple_page(
  softui::box(width = 5,
              title = "Progress bars",
              progressbar(20, status = "success", style = "margin-bottom: 30px;"),
              progressbar(32, status = "info", style = "margin-bottom: 30px;"),
              progressbar(49, status = "warning", style = "margin-bottom: 30px;"),
              progressbar(60, status = "danger", style = "margin-bottom: 30px;")
  )
)

server <- function(input, output, session) {
  
}
if(interactive()){
  shinyApp(ui, server)  
}
}
