% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property_table.R
\name{property_table}
\alias{property_table}
\title{Formatted table of properties}
\usage{
property_table(
  data,
  full_width = TRUE,
  show_na = TRUE,
  bold_property = TRUE,
  css_property = "font-size: 1.1em;",
  css_value = "font-size: 1.0em;"
)
}
\arguments{
\item{data}{A dataframe with one row.}

\item{full_width}{If TRUE, fills the container 100\%}

\item{show_na}{if TRUE, NA values are shown. If FALSE, rows with a value NA are not shown}

\item{bold_property}{Whether the property column is in bold font (default)}

\item{css_property}{CSS for the 'property' column in the table}

\item{css_value}{CSS for the 'value' column in the table}
}
\description{
Each column in \code{data} will be a row in this table, with the column name
\strong{bold} in the left column, and the value in the right column.
}
\details{
Replaces old function \code{bold_table_row} in use in many applications.
}
\examples{

# Use tibble() instead of data.frame() so the names do not get changed:
property_table(tibble::tibble(Kenmerk = "groot huis", Postcode = "123AB", `Aantal inwoners` = 2),
full_width = FALSE)
}
