% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactafel.R
\name{reactafel}
\alias{reactafel}
\title{Reactable, with NL and softui / Bootstrap 5.1 specific settings.}
\usage{
reactafel(
  ...,
  wrap = FALSE,
  striped = TRUE,
  resizable = TRUE,
  pageLength = 10,
  rownames = FALSE,
  ordering = TRUE
)
}
\arguments{
\item{\dots}{Further arguments to \code{reactable}, especially a dataframe as first argument (unnamed).}

\item{pageLength}{Number of rows to show}

\item{rownames}{Default FALSE, suppresses row names.}

\item{ordering}{Default TRUE, whether to allow column ordering}
}
\description{
Reactable, with NL and softui / Bootstrap 5.1 specific settings.
}
