% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactafel_click_handler.R
\name{reactafel_click_handler}
\alias{reactafel_click_handler}
\title{Reactable click handler}
\usage{
reactafel_click_handler(id = NULL, what = "data", session = NULL)
}
\arguments{
\item{id}{id that is used to catch click event}
}
\description{
Reactable click handler
}
