% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_application_info.R
\name{read_application_info}
\alias{read_application_info}
\title{Read application info (this_version or rsconnect deploy log)}
\usage{
read_application_info(read_git = TRUE)
}
\description{
Reads the content of  files "this_version.yml" and "shintoconnect_manifest.yml",
if they exist, and returns the contents as a list with components rsconnect and this_version.
Either of those is NULL if either of the files don't exist.
}
\seealso{
\code{\link{appInfoModule}}, \code{\link{make_deploy_project}}
}
