% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_news_modal.R
\name{read_application_news}
\alias{read_application_news}
\title{Read the NEWS.md file and maybe filter by category (tenant)}
\usage{
read_application_news(
  category = NULL,
  min_version = NULL,
  news_file = "NEWS.md"
)
}
\description{
Read the NEWS.md file and maybe filter by category (tenant)
}
