% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_git_version.R
\name{read_git_version}
\alias{read_git_version}
\title{Read most recent git SHA for the current working directory}
\usage{
read_git_version()
}
\description{
Read most recent git SHA for the current working directory
}
\details{
Returns a list with components 'sha' (the first 6 chars of the most recent
git commit SHA) and the date of last update. If the command fails (for example, this is
not a git repos), it returns "unknown" for both.
}
