% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_background_grey_level.R
\name{set_background_grey_level}
\alias{set_background_grey_level}
\title{Set body background grey level for higher contrast}
\usage{
set_background_grey_level(level, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{level}{100, 200, 300, etc.}
}
\description{
Set body background grey level for higher contrast
}
