% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_css_variable.R
\name{set_css_variable}
\alias{set_css_variable}
\title{Set a root: CSS variable for softui}
\usage{
set_css_variable(variable, value, session = shiny::getDefaultReactiveDomain())
}
\description{
Set a root: CSS variable for softui
}
