% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_sidebar.R
\name{dashboard_sidebar}
\alias{dashboard_sidebar}
\alias{sidebar_menu}
\alias{menu_item}
\alias{menu_link}
\title{Dashboard sidebar}
\usage{
dashboard_sidebar(arg)

sidebar_menu(..., id = NULL, .list = NULL)

menu_item(text, ..., icon = bsicon("bar-chart-fill"), tabName = NULL)

menu_link(tabName, icon, text)
}
\arguments{
\item{\dots}{`menu_item`'s}

\item{id}{Id for the sidebar, currently unused}

\item{.list}{Further `menu_item`'s can be passed as a list instead of in \dots.}
}
\description{
Dashboard sidebar

Sidebar menu

Sidebar menu item
}
