% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{softui_dependencies}
\alias{softui_dependencies}
\title{Softui Dependency for the UI}
\usage{
softui_dependencies()
}
\description{
When not using softui::dashboard_page, still load the dependencies.
Place this in your main UI : softui::softui_dependencies()
}
