% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_box.R
\name{sub_box}
\alias{sub_box}
\title{Box within a box}
\usage{
sub_box(..., title = NULL, grey_level = 0.1)
}
\arguments{
\item{title}{A title for the box}

\item{grey_level}{Value between 0.1 (standard) and 0.9 (very dark), in steps of 0.1.}
}
\description{
You are not allowed to place a `softui::box` inside another `softui::box`, but you can
use this `softui::sub_box`
}
