% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_box.R
\name{tab_panel}
\alias{tab_panel}
\alias{tab_box}
\title{Tab panels}
\usage{
tab_panel(title = "", ..., value = title, icon = NULL)

tab_box(
  ...,
  selected = NULL,
  width = 12,
  type = c("tabs", "pills"),
  fill = FALSE
)
}
\description{
Tab panels
}
