% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabset_panel.R
\name{tabset_panel}
\alias{tabset_panel}
\title{Tab panel}
\usage{
tabset_panel(
  ...,
  id = NULL,
  selected = NULL,
  width = 12,
  fill = FALSE,
  type = c("pills", "tabs"),
  style = ""
)
}
\description{
Tab panel
}
