% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_news_modal.R
\name{tenantNewsModuleUI}
\alias{tenantNewsModuleUI}
\alias{tenantNewsModule}
\title{Shiny module to show all software news for a tenant}
\usage{
tenantNewsModuleUI(id)

tenantNewsModule(input, output, session, category = NULL)
}
\description{
Shiny module to show all software news for a tenant
}
