% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tabpanel.R
\name{update_tabpanel}
\alias{update_tabpanel}
\title{Update tab_box}
\usage{
update_tabpanel(id, selected, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{id of the tab_box (required)}

\item{selected}{Which tab to select (using 'value' of the \code{tab_panel})}

\item{session}{Not required}
}
\description{
Simulates a click on a tab inside a \code{softui::tab_box} (replaces updateTabsetpanel)
}
