% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{valid_statuses}
\alias{valid_statuses}
\title{Valid statuses}
\description{
These status strings correspond to colors as defined in Bootstrap's CSS.
Although the colors can vary depending on the particular CSS selector, they
generally appear as follows:
}
\details{
\itemize{
\item \code{primary} Dark blue
\item \code{secondary} Grey / blue
\item \code{success} Green
\item \code{info} Lighter blue
\item \code{warning} Orange
\item \code{danger} Red
\item \code{light} Light grey
\item \code{dark} Dark grey / black
}
}
\keyword{internal}
