% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box.R
\name{value_box}
\alias{value_box}
\title{Value box}
\usage{
value_box(
  value,
  title,
  sub_value = "",
  sub_status = "success",
  icon = icon("chart-bar"),
  width = 4,
  status = "primary",
  format_value = FALSE,
  height = NULL
)
}
\arguments{
\item{value}{A value to display}

\item{sub_value}{A subtitle under the value}

\item{sub_status}{Status color of the badge (best: "success" for positive, "danger" for negative, or see \code{?valid_statuses})}

\item{icon}{An icon to place}

\item{format_value}{If TRUE, values like "1000" are formatted as "1,000". For this to work, value has to be numeric!}

\item{height}{Height of the value box (NULL to autosize) (400 or "400px")}
}
\description{
Value box
}
\details{
Makes a softui 'value box' with icon, value, title, and badge (e.g. +12, -80) with a status color.
Use in a \code{uiOutput}.
}
