% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_icon_list.R
\name{value_icon_table}
\alias{value_icon_table}
\alias{value_icon_item}
\title{Value icon table for use in a box}
\usage{
value_icon_table(...)

value_icon_item(
  title = NULL,
  icon = NULL,
  value = 0,
  value_suffix = "items",
  sub_value = NULL,
  sub_value_suffix = NULL
)
}
\description{
Value icon table for use in a box
}
