% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_select_input.R
\name{virtual_select_input}
\alias{virtual_select_input}
\title{Simple wrapper around virtualSelectInput}
\usage{
virtual_select_input(
  inputId,
  label,
  choices,
  search = TRUE,
  hideClearButton = FALSE,
  placeholder = "Maak selectie ...",
  noOptionsText = "Geen keuze mogelijk",
  noSearchResultsText = "Geen resultaten",
  selectAllText = "Selecteer alles",
  searchPlaceholderText = "Zoeken ...",
  optionsSelectedText = "Geselecteerd",
  allOptionsSelectedText = "Alles geselecteerd",
  zIndex = 10,
  ...
)
}
\arguments{
\item{inputId}{Shiny input ID}

\item{label}{Label above the select input}

\item{choices}{Vector or list, see [shinywidgets::prepare_choices()] for 
more advanced setting of choices in this select input (descriptions, grouping).}

\item{search}{If TRUE, adds a search menu above the choices (default TRUE)}

\item{\dots}{More arguments to [shinywidgets::virtualSelectInput()]}
}
\description{
Wrapper around [shinywidgets::virtualSelectInput()] with Dutch 
language settings and a better default for the z-index.
}
