% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_tooltip.R
\name{with_tooltip}
\alias{with_tooltip}
\title{Add a tooltip next to a label}
\usage{
with_tooltip(
  label = NULL,
  tooltip,
  icon = bsicon("info-circle-fill"),
  position = "top",
  ...
)
}
\arguments{
\item{label}{Label, the text to appear left of the icon. If you want only the icon, do not set.}

\item{tooltip}{Text to show when hovering over the icon.}

\item{icon}{The icon to show next to the label}

\item{position}{Position of the tooltip, default 'top' (see `?prompter::add_prompt` for options)}

\item{\dots}{Further arguments passed to [prompter::add_prompt()]}
}
\description{
Adds an icon right of a label, on hover shows help text. 
Uses the `prompter` package. For this to work, add `use_prompter()` in the UI function of
your app, this is already loaded by `dashboard_page` and `simple_page`.
}
