% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_button.R
\name{action_button}
\alias{action_button}
\title{Simple wrapper around actionButton with softui CSS classes}
\usage{
action_button(..., class = "", status = NULL)
}
\arguments{
\item{\dots}{Arguments sent to \code{shiny::actionButton}}

\item{class}{Extra CSS classes}

\item{status}{Status color (see \code{?valid_statuses})}
}
\description{
Simple wrapper around actionButton with softui CSS classes
}
