% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box}
\alias{box}
\title{Content box}
\usage{
box(
  ...,
  width = 12,
  title = NULL,
  subtitle = NULL,
  tag = "",
  collapsible = TRUE,
  collapsed = FALSE,
  header_ui = NULL
)
}
\arguments{
\item{width}{Bootstrap width}

\item{title}{Title for the box}

\item{subtitle}{Subtitle for the box}

\item{tag}{Short label text above the title (in CAPS)}

\item{header_ui}{More UI to place just under the title}
}
\description{
Content box
}
