% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_collection.R
\name{box_panel}
\alias{box_panel}
\alias{box_collection}
\title{Indexed boxes}
\usage{
box_panel(
  title = "",
  ...,
  value = title,
  icon = NULL,
  class = "",
  width = NULL
)

box_collection(
  ...,
  id = NULL,
  selected = NULL,
  title = NULL,
  width_nav = 3,
  width_boxes = 9,
  fill = FALSE,
  style = "",
  .list = NULL
)
}
\arguments{
\item{width_nav}{Width of the menu on the left (bootstrap units)}

\item{width_boxes}{Width of the boxes on the right}
}
\description{
Works like a tab_box, except a menu appears on the left (in a box),
and content appears in boxes on the right.
}
\examples{
 ui <- softui::simple_page(
box_collection(

 box_panel(title = "Box 1", tags$p("Dit is de eerste box met content")),
 box_panel(title = "Box 2", tags$p("Dit is de tweede box met content")),
 box_panel(title = "Box 3", tags$p("Dit is de derde box met content")),
 box_panel(title = "Box 4", tags$p("Dit is de vierde box met content"))
 
)

)

shinyApp(ui, function(input,output){})
}
