% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_items.R
\name{dashboard_body}
\alias{dashboard_body}
\alias{tab_items}
\alias{tab_item}
\title{Dashboard body}
\usage{
dashboard_body(...)

tab_items(..., .list = NULL)

tab_item(tabName, ...)
}
\arguments{
\item{\dots}{A `tab_items` object, and maybe other HTML.}
}
\description{
Makes the dashboard body of a dashboard page. See `?dashboard_page` for a minimal example.
Accepts only a `tab_items` object, which has multiple `tab_item`'s in it, each with the UI for a 'page' 
(typically a shiny module UI function).
}
