% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_header.R
\name{dashboard_header}
\alias{dashboard_header}
\alias{header_item}
\alias{dropdown_item}
\title{The dashboard header}
\usage{
dashboard_header(..., tag_line = NULL, left_content = NULL, info = NULL)

header_item(...)

dropdown_item(icon, ..., header = NULL, align_right = FALSE)
}
\arguments{
\item{\dots}{Any header items, usually made with \code{softui::dropdown_item}.}

\item{tag_line}{A short statement to appear in the header (italic text)}

\item{left_content}{UI to place on the left (just right of the tagline). Untested.}

\item{info}{Passed to appInfoUI()}
}
\description{
The dashboard header

UI to place in the header bar (to the right)

Dropdown item in the header bar
}
