% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_page.R
\name{dashboard_page}
\alias{dashboard_page}
\title{Soft UI dashboard page}
\usage{
dashboard_page(
  header,
  sidebar,
  body,
  title = "",
  icon = softui::bsicon("speedometer"),
  loadingscreen_time = 4,
  busybar_color = "#346E6E",
  shinto_logo = TRUE,
  disconnect_message = "De sessie is verlopen of er ging iets fout!"
)
}
\arguments{
\item{header}{Made with \code{softui::dashboard_header}}

\item{sidebar}{Made with \code{softui::dashboard_sidebar}}

\item{body}{Made with \code{softui::dashboard_body}}

\item{title}{App title (shown topleft)}

\item{icon}{An icon for the very topleft}

\item{loadingscreen_time}{How many seconds to show the Shinto loading screen}

\item{busybar_color}{Color of the 'busy bar' when shiny is busy}

\item{disconnect_message}{Message to show when session is disconnected (timeout / error)}
}
\description{
Main function to make the Soft UI dashboard page in shiny.
}
\examples{
sidebar <- softui::dashboard_sidebar(
  softui::sidebar_menu(
    softui::menu_item("Page1", tabName = "page1", icon = bsicon("bar-chart-fill")),
    softui::menu_item("Page2", tabName = "page2", icon = bsicon("tools"))
  )
)

body <- softui::dashboard_body(
  
  softui::tab_items(
    softui::tab_item("page1",
                     softui::box(title = "Page 1", width = 12)
    ),
    
    softui::tab_item("page2",
                     softui::box(title = "Page 2", width = 12)
    )
  )
)

header <- softui::dashboard_header(
  tag_line = "A tag line"
)

ui <- softui::dashboard_page(title = "Shinto App",
                             loadingscreen_time = 1,
                             header = header,
                             sidebar = sidebar,
                             body = body)

server <- function(input, output, session) {}
}
