% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_page.R
\name{dashboard_page}
\alias{dashboard_page}
\title{Soft UI dashboard page}
\usage{
dashboard_page(header, sidebar, body, title = "", icon = bsicon("speedometer"))
}
\description{
Soft UI dashboard page
}
