% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatafel.R
\name{datatafel}
\alias{datatafel}
\title{Datatable, with NL and softui / Bootstrap 5.1 specific settings.}
\usage{
datatafel(
  ...,
  dom = "frtip",
  pageLength = 10,
  escape = FALSE,
  rownames = FALSE,
  ordering = TRUE,
  extensions = "Responsive",
  selection = "single",
  scrollX = FALSE,
  options = list(),
  lengthChange = FALSE,
  default_style = "bootstrap4",
  class = "row-border hover"
)
}
\arguments{
\item{\dots}{Further arguments to \code{datatable}, especially a dataframe as first argument (unnamed).}

\item{dom}{DOM elements for the datatable ("t" for just the table)}

\item{pageLength}{Number of rows to show}

\item{escape}{Default TRUE, allow HTML content in the table}

\item{rownames}{Default FALSE, suppresses row names.}

\item{ordering}{Default TRUE, whether to allow column ordering}

\item{extensions}{Default 'Responsive' to show long rows in a collapsible.}

\item{selection}{Rows to select (single or multiple)}

\item{scrollX}{If TRUE, allows scrolling for very wide tables (then set extensions = NULL!)}

\item{options}{Passed to datatable \code{options} argument}

\item{class}{Datatable CSS classes.}
}
\description{
Datatable, with NL and softui / Bootstrap 5.1 specific settings.
}
