% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconActionButton.R
\name{iconActionButton}
\alias{iconActionButton}
\title{Icon only button}
\usage{
iconActionButton(id, icon, status = "success", class = "", ...)
}
\arguments{
\item{id}{The shiny input id}

\item{icon}{A `softui::bsicon` or `shiny::icon`.}

\item{status}{The status color (see `?valid_statuses`).}

\item{\dots}{Additional arguments to `shiny::actionButton`}
}
\description{
A 'softui' round button with only an icon. Simple wrapper around `shiny::actionButton`. 
Uses various softui CSS classes for styling.
}
\examples{

ui <- softui::simple_page(
  
  softui::box(width = 4,
              title = "iconActionButton",
              
              iconActionButton("id2", icon = bsicon("exclamation-circle-fill"), 
                               status = "primary"),
              shiny::tags$br(),
              iconActionButton("id1", icon = bsicon("award-fill"), status = "success"),
              shiny::tags$br(),
              iconActionButton("id2", icon = bsicon("stop-circle-fill"), 
                               status = "danger", class = "btn-lg")
  )
  
  
)
shinyApp(ui,server = function(input,output){})
}
