% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modaldeps.R
\name{modal}
\alias{modal}
\title{Convenient wrapper around modalDialog}
\usage{
modal(
  ...,
  title = "",
  easyClose = TRUE,
  confirm_button = TRUE,
  id_confirm = "btn_confirm_modal",
  confirm_txt = "OK",
  confirm_icon = softui::bsicon("check-lg"),
  close_button = TRUE,
  close_txt = "Sluiten",
  close_icon = softui::bsicon("x-lg"),
  remove_modal_on_confirm = TRUE,
  size = c("m", "s", "l", "xl", "fullscreen"),
  fullscreen_up_to = c("always", "sm", "md", "lg", "xl", "xxl"),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{title}{Title for the modal}

\item{easyClose}{Can the modal close by clicking outside the modal?}

\item{id_confirm}{Id of the 'OK' button, will be namespaced!}

\item{confirm_txt}{Text on 'OK' button}

\item{confirm_icon}{Icon on 'OK' button}

\item{close_txt}{Text on 'Close' button}

\item{close_icon}{Icon on 'Close' icon}

\item{remove_modal_on_confirm}{If TRUE, always removes the modal on OK click. 
Otherwise, call `removeModal` yourself. This may be needed when some input 
validation is done before the modal can be removed.}

\item{session}{Shiny session object (no need to set)}
}
\description{
Convenient wrapper around modalDialog
}
