% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_link.R
\name{page_link}
\alias{page_link}
\alias{route_page_link}
\title{Utilities for links to pages}
\usage{
page_link(
  id,
  label,
  id_out = "click_id",
  session = shiny::getDefaultReactiveDomain(),
  class = "page_link"
)

route_page_link(
  id,
  tabName,
  input,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Value of the ID that will be sent as link information (e.g. person ID, location ID)}

\item{label}{Label to appear in the link}

\item{id_out}{Name of the input ID to set when clicked on a link (namespacing in modules is done automatically)}

\item{session}{Shiny session object (no need to set)}

\item{class}{CSS class of the text in the link}

\item{tabName}{Which page to jump to on clicking the link}

\item{input}{Shiny input object}
}
\description{
Utilities for links to pages
}
