% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictem_menu_item.R
\name{restricted_menu_item}
\alias{restricted_menu_item}
\alias{restricted_menu_item_module}
\title{Menu item with restricted access}
\usage{
restricted_menu_item(id)

restricted_menu_item_module(
  input,
  output,
  session,
  text,
  tabName,
  icon,
  user_group,
  allowed_groups
)
}
\description{
Menu item with restricted access
}
