% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_sidebar.R
\name{dashboard_sidebar}
\alias{dashboard_sidebar}
\alias{sidebar_menu}
\alias{menu_item}
\title{Dashboard sidebar}
\usage{
dashboard_sidebar(arg)

sidebar_menu(..., id = NULL, .list = NULL)

menu_item(text, ..., icon = bsicon("bar-chart-fill"), tabName = NULL)
}
\description{
Dashboard sidebar

Sidebar menu

Sidebar menu item
}
