% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_page.R
\name{simple_page}
\alias{simple_page}
\title{Simple page without sidebar or header}
\usage{
simple_page(...)
}
\description{
Make a simple page with the correct background color, and all of `softui`'s dependencies.
Useful for testing or simple applications
}
