% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_box.R
\name{sub_box}
\alias{sub_box}
\title{Box within a box}
\usage{
sub_box(
  ...,
  title = NULL,
  icon = NULL,
  grey_level = 0.1,
  collapsible = TRUE,
  collapsed = FALSE,
  margin_bottom = 0
)
}
\arguments{
\item{\dots}{Any HTML (or text) items to appear in the box.}

\item{title}{A title for the box}

\item{icon}{An icon (optional), placed left of the title}

\item{grey_level}{Value between 0.1 (standard) and 0.9 (very dark), in steps of 0.1.}

\item{collapsible}{If TRUE (default), can be minimized}

\item{collapsed}{If TRUE, starts minimized}

\item{margin_bottom}{Margin below the \code{sub_box}, in pixels}
}
\description{
You are not allowed to place a \code{softui::box} inside another \code{softui::box}, but you can
use this \code{sub_box}, or \code{?alert_box} for a message-type box.
}
