% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_box.R
\name{tab_panel}
\alias{tab_panel}
\alias{tab_box}
\title{Tab panels}
\usage{
tab_panel(title = "", ..., value = title, icon = NULL, style = "")

tab_box(
  ...,
  id = NULL,
  selected = NULL,
  width = 12,
  fill = FALSE,
  ui_above_nav = NULL,
  ui_below_nav = NULL,
  style = ""
)
}
\arguments{
\item{title}{Title for the tab}

\item{\dots}{Content for the tab}

\item{value}{Name of the tab to refer to when using `softui::update_tabpanel` (defaults to title)}

\item{icon}{An icon, with `shiny::icon` or `softui::bsicon`}

\item{style}{Extra CSS for the `card` class.}

\item{id}{Optional id for the tab_box}

\item{selected}{Which tab is selected initially? Defaults to the first tab}

\item{width}{Width of tab_box in bootstrap units (default = 12)}

\item{fill}{If TRUE, makes the tab panel titles fill the entire available width}

\item{ui_above_nav}{Extra UI to place in the box *above* the nav links (i.e. tab names)}

\item{ui_below_nav}{Extra UI to place below the nav links, this is UI that will appear at the top
of each tab (so you don't have to repeat identical content within each `tab_panel` in `tab_box`).}
}
\description{
Replaces `shiny::tabPanel`, and is used only for placing tabs inside a `softui::tab_box`
}
\details{
More details here!
}
