% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_box.R
\name{tab_panel}
\alias{tab_panel}
\alias{tab_box}
\title{Tab panels}
\usage{
tab_panel(title = "", ..., value = title, icon = NULL, style = "")

tab_box(..., id = NULL, selected = NULL, width = 12, fill = FALSE)
}
\arguments{
\item{id}{Optional id for the tab_box}
}
\description{
Tab panels
}
