% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline}
\alias{timeline}
\alias{timeline_block}
\title{Softui Timeline}
\usage{
timeline(...)

timeline_block(
  title,
  timestamp,
  text = "",
  icon_name = "bell-fill",
  icon_status = c("primary", "secondary", "info", "success", "danger", "warning",
    "light", "dark"),
  badge_text = NULL,
  badge_status = c("primary", "secondary", "info", "success", "danger", "warning",
    "light", "dark")
)
}
\description{
Softui Timeline

Softui Timeline
}
