% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline}
\alias{timeline}
\alias{timeline_block}
\alias{timeline_from_data}
\title{Softui Timeline}
\usage{
timeline(..., .list = NULL)

timeline_block(
  title,
  timestamp,
  text = "",
  icon_name = "bell-fill",
  icon_status = "primary",
  badge_text = NULL,
  badge_status = "success"
)

timeline_from_data(data)
}
\arguments{
\item{title}{Title for the timeline block}

\item{timestamp}{Date or time for the block (format yourself!)}

\item{text}{More text to describe the block}

\item{icon_name}{Name of the \code{bsicon} for this block (character)}

\item{icon_status}{color of the icon (see \code{?valid_statuses})}

\item{badge_text}{Text for a badge to appear at the bottom of the timeline block}

\item{badge_status}{Status color for the badge (see \code{?valid_statuses})}

\item{data}{A dataframe with columns: "timestamp","title","text","icon_name","icon_status"}
}
\description{
Softui Timeline

Softui Timeline
}
\details{
The \code{timeline} function makes a timeline consisting of \code{timeline_block}'s.
Use the \code{timeline_from_data} function to make a timeline from a dataframe. Can be used in a \code{uiOutput}.
}
\examples{

# Manual timeline
# (only works in a softui page!)
timeline(
  timeline_block("Geboorte", "1956-6-1", icon_name = "bell-fill", icon_status = "success",
               text = shinipsum::random_text(nwords=20)),
  timeline_block("Gebeurtenis 1", "1970-1-6", icon_name = "bag-fill", icon_status = "info",
               text = shinipsum::random_text(nwords=20)),
  timeline_block("Gebeurtenis 2", "2002-6-1", icon_name = "bandaid-fill", icon_status = "warning",
               text = shinipsum::random_text(nwords=20)),
  timeline_block("Gebeurtenis 3", "2014-7-15", icon_name = "chat-fill", icon_status = "primary",
               text = shinipsum::random_text(nwords=20)),
  timeline_block("Overlijden", "2021-4-1", icon_name = "hourglass-bottom", icon_status = "danger",
               text = shinipsum::random_text(nwords=20))
)

# Timeline from dataframe
time_data <- tibble::tribble(
   ~timestamp, ~title, ~text, ~icon_name, ~icon_status, 
   "1956-6-1","Geboorte", shinipsum::random_text(nwords=20),"bell-fill","success",
   "1970-1-6","Gebeurtenis 1",shinipsum::random_text(nwords=20),"bag-fill", "info",
   "2002-6-1","Gebeurtenis 2",shinipsum::random_text(nwords=20),"bandaid-fill", "warning",
   "2014-7-15","Gebeurtenis 3",shinipsum::random_text(nwords=20),"chat-fill", "primary",
   "2021-4-1", "Overlijden",shinipsum::random_text(nwords=20), "hourglass-bottom", "danger"
 )
 timeline_from_data(time_data)
}
