% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box.R
\name{value_box}
\alias{value_box}
\title{Value box}
\usage{
value_box(
  value,
  title,
  sub_value = "",
  sub_status = c("success", "danger"),
  icon = icon("chart-bar"),
  width = 4,
  height = NULL
)
}
\arguments{
\item{value}{A value to display}

\item{sub_value}{A subtitle under the value}

\item{sub_status}{"success" or "danger" to color the subtitle}

\item{icon}{An icon to place}

\item{height}{Height of the value box (NULL to autosize) (400 or "400px")}
}
\description{
Value box
}
